requirements.txt
fast-mcp==0.1.2
requests==2.32.3
pydantic==2.8.2
uvicorn==0.30.1
python-dotenv==1.0.1